feat:Add authentication events handler for auditory and separate project security config for prod
